refactor(selector4nix,tests/integration): move `Url` from `Substituter` to a common module
