+++ src/application/substituter/actor/mod.rs
@@ -4,6 +4,6 @@ pub use runner::{SubstituterActor, SubstituterRequest};
 
 use selector4nix_actor::registry::Registry;
 
-use crate::domain::substituter::model::Url;
+use crate::domain::common::url::Url;
 
 pub type SubstituterActorRegistry = Registry<Url, SubstituterActor>;
