+++ src/domain/mod.rs
@@ -1,3 +1,4 @@
+pub mod common;
 pub mod nar_file;
 pub mod nar_info;
 pub mod substituter;
