+++ src/domain/nar_info/model/proxy_nar_info_data.rs
@@ -1,7 +1,8 @@
 use getset::Getters;
 
+use crate::domain::common::url::Url;
 use crate::domain::nar_info::model::{NarFileName, UpstreamNarInfoData};
-use crate::domain::substituter::model::{SubstituterMeta, Url};
+use crate::domain::substituter::model::SubstituterMeta;
 
 #[derive(Debug, Clone, PartialEq, Eq, Hash, Getters)]
 #[getset(get = "pub")]
