+++ src/domain/nar_info/model/upstream_nar_info_data.rs
@@ -1,8 +1,8 @@
 use getset::Getters;
 use snafu::{OptionExt, ResultExt, Snafu};
 
+use crate::domain::common::url::Url;
 use crate::domain::nar_info::model::{NarFileName, TryNewNarFileNameError};
-use crate::domain::substituter::model::Url;
 
 #[derive(Debug, Clone, PartialEq, Eq, Hash, Getters)]
 pub struct UpstreamNarInfoData {
