+++ src/domain/substituter/model/mod.rs
@@ -2,10 +2,8 @@ mod availability;
 mod priority;
 mod substituter;
 mod substituter_meta;
-mod url;
 
 pub use availability::Availability;
 pub use priority::{Priority, TryNewPriorityError};
 pub use substituter::{PeriodicProbingOption, ProbedState, Substituter, UpdateSubstituterEvent};
 pub use substituter_meta::SubstituterMeta;
-pub use url::{TryNewUrlError, Url};
