+++ tests/integration/mock/nar_info_provider.rs
@@ -2,9 +2,9 @@ use std::collections::HashMap;
 use std::time::Duration;
 
 use async_trait::async_trait;
+use selector4nix::domain::common::url::Url;
 use selector4nix::domain::nar_info::port::error_ctx::{OfflineSnafu, ServiceSnafu};
 use selector4nix::domain::nar_info::port::{NarInfoProvider, NarInfoQueryData, QueryNarInfoError};
-use selector4nix::domain::substituter::model::Url;
 use snafu::ResultExt;
 
 #[derive(Debug, Clone, PartialEq, Eq)]
