+++ tests/integration/nar_info_service_test.rs
@@ -2,11 +2,12 @@ use std::collections::HashSet;
 use std::sync::Arc;
 use std::time::Duration;
 
+use selector4nix::domain::common::url::Url;
 use selector4nix::domain::nar_info::model::{NarUrlRewriteOption, StorePathHash};
 use selector4nix::domain::nar_info::port::NarInfoQueryData;
 use selector4nix::domain::nar_info::{NarInfoService, ResolveNarInfoError, ResolveNarInfoEvent};
 use selector4nix::domain::substituter::SubstituterRepository;
-use selector4nix::domain::substituter::model::{Substituter, Url};
+use selector4nix::domain::substituter::model::Substituter;
 use selector4nix::infrastructure::repository::InMemorySubstituterRepository;
 
 use crate::fixture::{nar_file, nar_info, substituter};
